added rgb raw image subscriber
